+++ dst/jobs/triggered/reportFlowTelemetry/run.js
@@ -1,0 +1,17 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+/**
+ * 測定データ報告
+ */
+const createDebug = require("debug");
+const Controller = require("../../../controller/reportFlowTelemetry");
+const debug = createDebug('sskts-monitoring-jobs');
+Controller.main()
+    .then(() => {
+    debug('success!');
+})
+    .catch((err) => {
+    // tslint:disable-next-line:no-console
+    console.error(err);
+    process.exit(1);
+});