+++ test/index-test.js
@@ -1,0 +1,6 @@
+// tslint:disable:no-useless-files
+// tslint:disable-next-line:no-suspicious-comment
+/**
+ * 測定データ報告テスト
+ * TODO テストコード不備
+ */